UI: Split registers and handler views
