+++ ui/src/app.rs
@@ -3,9 +3,10 @@ use emu::{cartridge_header::CartridgeHeader, gba::Gba};
 use logger::log;
 use std::io::Read;
 
-use super::cpu_inspector::CpuInspector;
+use super::cpu_registers::CpuRegisters;
 use crate::{
-    about, gba_display::GbaDisplay, palette_visualizer::PaletteVisualizer, ui_traits::UiTool,
+    about, cpu_handler::CpuHandler, gba_display::GbaDisplay, palette_visualizer::PaletteVisualizer,
+    ui_traits::UiTool,
 };
 
 use std::{
@@ -50,7 +51,8 @@ impl ClementineApp {
 
         Self::from_tools(vec![
             Box::<about::About>::default(),
-            Box::new(CpuInspector::new(Arc::clone(&arc_gba))),
+            Box::new(CpuRegisters::new(Arc::clone(&arc_gba))),
+            Box::new(CpuHandler::new(Arc::clone(&arc_gba))),
             Box::new(GbaDisplay::new(Arc::clone(&arc_gba))),
             Box::new(PaletteVisualizer::new(arc_gba)),
             Box::new(disassembler),
